Rename GuestFields for consistency
